@@ -3,9 +3,9 @@ library color;
 import 'package:flutter/material.dart';
 
 class ColorPalettes {
-  Color lightPrimary = Color(0xfffcfcff);
-  Color primaryDarkColor = Colors.blue[900];
-  Color primaryColor = Colors.lightBlue[800];
-  Color accentColor = Colors.yellow[500];
-  Color errorColor = Color(0xffe11e27);
+  Color? lightPrimary = Color(0xfffcfcff);
+  Color? primaryDarkColor = Colors.blue[900];
+  Color? primaryColor = Colors.lightBlue[800];
+  Color? accentColor = Colors.yellow[500];
+  Color? errorColor = Color(0xffe11e27);
 }
